cleanup and javadoc for ba
